Set project version for PROJECT_NUMBER in `doxygen` task (#11)
The default `doxygen` task attempted to set the version too early.
unfortunately in the current architecture there is no setup for lazy
evaluation and attempteing to do so would take it beyond the scope of a
simple patch release.

The fix is to inspect the doxygenProperties at a point when setting the
defaults. This will only occur at execution  time, which is late enough
in the process.